refactor scripts to use new config and update readme

+++ trace_validation_pipeline.py
@@ -4,39 +4,52 @@ import run_impl
 import trace_merger
 import tla_trace_validation
 import argparse
+import ndjson
 
-parser = argparse.ArgumentParser(
-    prog='Trace validation pipeline',
-    description='This program aims to execute a pipeline that validate implementation of KeyValueStore against a formal spec.')
+def read_json(filename):
+    with open(filename) as f:
+        return ndjson.load(f)
 
+def get_files(config):
+    files = []
+    for line in config:  
+        for v in line.values():
+            files += [f+".ndjson" for f in v]
+    return files
+
+parser = argparse.ArgumentParser("")
 parser.add_argument('-c', '--compile', type=bool, action=argparse.BooleanOptionalAction)
-
+parser.add_argument('--config', type=str, required=False, default="twophase.ndjson.conf", help="Config file")
 args = parser.parse_args()
 
+config = read_json(args.config)
+files = get_files(config)
+
+# Clean up
 print("# Clean up")
-
-trace_files = [f for f in os.listdir(".") if f.endswith('.ndjson')]
-print(f"Cleanup: {trace_files}")
+trace_files = files + ["trace-tla.ndjson"]
+print(f"Cleanup: {files}")
 for trace_file in trace_files:
-    os.remove(trace_file)
+    if os.path.isfile(trace_file):
+        os.remove(trace_file)
 
+# Compile
 if args.compile:
     print("# Package.\n")
     subprocess.run(["mvn", "package"])
 
+# Run
 print("# Start implementation.\n")
+rms = config[0]["RM"]
+tm = config[1]["TM"][0]
+run_impl.run(rms,tm)
 
-run_impl.run()
-
+# Merge traces 
 print("# Merge trace with config.\n")
+trace_merger.run(files,sort=True, remove_meta=True, out="trace-tla.ndjson")
 
-trace_tla = trace_merger.run(["."], config="twophase.ndjson.conf", sort=True, remove_meta=True)
-# Write to file
-with open("trace-tla.ndjson", "w") as f:
-    f.write(trace_tla)
-
+# Validate trace
 print("# Start TLA+ trace spec.\n")
-
 tla_trace_validation.run_tla("spec/TwoPhaseTrace.tla","trace-tla.ndjson","twophase.ndjson.conf")
 
-print("End pipeline.")
+# print("End pipeline.")
